Use Custom NotFoundException

+++ Store.Application/Features/Attachment/Handlers/Commands/DeleteAttachmentRequestHandler.cs
@@ -1,7 +1,9 @@
 ﻿using AutoMapper;
 using MediatR;
+using Store.Application.Exceptions;
 using Store.Application.Features.Attachment.Requests.Command;
 using Store.Application.Persistence.Contracts;
+using Store.Domain;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -23,6 +25,10 @@ namespace Store.Application.Features.Attachment.Handlers.Commands
         public async Task Handle(DeleteAttachmentRequest request, CancellationToken cancellationToken)
         {
             var attachment = await attachmentRepository.Get(request.Id);
+            if (attachment == null)
+            {
+                throw new NotFoundException(nameof(Domain.Attachment), request.Id);
+            }
             await attachmentRepository.Delete(attachment);
         }
     }
