+++ Store.Application/Features/OrderDetail/Handlers/Commands/DeleteOrderDetailCommandRequestHandler.cs
@@ -1,7 +1,9 @@
 ﻿using AutoMapper;
 using MediatR;
+using Store.Application.Exceptions;
 using Store.Application.Features.OrderDetail.Requests.Commands;
 using Store.Application.Persistence.Contracts;
+using Store.Domain;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -23,6 +25,10 @@ namespace Store.Application.Features.OrderDetail.Handlers.Commands
         public async Task Handle(DeleteOrderDetailCommandRequest request, CancellationToken cancellationToken)
         {
             var orderDetail =await orderDetailRepository.Get(request.Id);
+            if (orderDetail == null)
+            {
+                throw new NotFoundException(nameof(Domain.OrderDetail), request.Id);
+            }
             await orderDetailRepository.Delete(orderDetail);
         }
     }
