+++ Store.Application/Features/Person/Handlers/Commands/DeletePersonCommandRequstHandler.cs
@@ -1,7 +1,9 @@
 ﻿using AutoMapper;
 using MediatR;
+using Store.Application.Exceptions;
 using Store.Application.Features.Person.Requests.Commands;
 using Store.Application.Persistence.Contracts;
+using Store.Domain;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -23,6 +25,10 @@ namespace Store.Application.Features.Person.Handlers.Commands
         public async Task Handle(DeletePersonCommandRequst request, CancellationToken cancellationToken)
         {
             var person=await personRepository.Get(request.Id);
+            if (person == null)
+            {
+                throw new NotFoundException(nameof(Domain.Person), request.Id);
+            }
             await personRepository.Delete(person);
         }
     }
